Move all benchmarks to benches/
