+++ benches/signature.rs
@@ -1,0 +1,12 @@
+#![feature(test)]
+extern crate solana;
+extern crate test;
+
+use solana::signature::GenKeys;
+use test::Bencher;
+
+#[bench]
+fn bench_gen_keys(b: &mut Bencher) {
+    let rnd = GenKeys::new([0u8; 32]);
+    b.iter(|| rnd.gen_n_keypairs(1000));
+}
